backend/requirements.txt
django
pandas
openpyxl
xlrd
djangorestframework
django-cors-headers
Babel
python-dotenv